requirements.txt
appdirs==1.4.3
asgiref==3.2.3
beautifulsoup4==4.8.2
bs4==0.0.1
cachetools==4.1.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
dialogflow==1.0.0
distlib==0.3.0
Django==3.0.3
django-bootstrap3==12.0.3
feedparser==5.2.1
filelock==3.0.12
Flask==1.1.2
fuzzywuzzy==0.18.0
gnewsclient==1.12
google-api-core==1.21.0
googleapis-common-protos==1.52.0
grpcio==1.30.0
gTTS==2.1.1
gTTS-token==1.1.3
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.5.0
importlib-resources==1.0.2
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.1.0
markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.2
packaging==20.3
panda==0.3.1
Pillow==7.0.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.8
pyglet==1.5.2
pymongo==3.10.1
pyparsing==2.4.6
PyQt5==5.14.1
PyQt5-sip==12.7.1
PyQtWebEngine==5.14.0
python-dateutil==2.8.1
python-dotenv==0.12.0
python-telegram-bot==12.8
pytz==2019.3
requests==2.22.0
reverse==0.1.0
rsa==4.6
simplejson==3.17.0
sip==5.1.1
six==1.14.0
soupsieve==1.9.5
SpeechRecognition==3.8.1
sqlparse==0.3.0
standalone==1.0.1
toml==0.10.0
urllib3==1.25.8
virtualenv==20.0.4
web.py==0.40
Werkzeug==1.0.1
zipp==3.0.0
